Splits user-events into more usable & well organized structure (#1914)
* To discuss

* To discuss

* Update page-tou.pug

* can i merge now

* maybe this will work?

* made changes as discussed with aem

* fixes some issues

* removes c word from filter

* Fixes formatting, adds other changes

* Update page-tou.pug

* fixes guessing in silent games + fix remake glitch

* allow line guesses in unlisted and casual games

* prevent @mod when a mod is seated

* Fixes mod pings again among other things, improves text at the end of the game

* file split, pausing in order to rebase meeman fix, lint fails but that's because i'm not done:

* forgot this

* file splits and fixes some crashes with forcevote

* bruh lint
