+++ routes/socket/user-events/flappy-hitler.js
@@ -1,0 +1,56 @@
+module.exports.handleFlappyEvent = (data, game) => {
+	if (!io.sockets.adapter.rooms[game.general.uid]) {
+		return;
+	}
+	const roomSockets = Object.keys(io.sockets.adapter.rooms[game.general.uid].sockets).map(sockedId => io.sockets.connected[sockedId]);
+	const updateFlappyRoom = newData => {
+		roomSockets.forEach(sock => {
+			if (sock) {
+				sock.emit('flappyUpdate', newData);
+			}
+		});
+	};
+
+	updateFlappyRoom(data);
+
+	if (data.type === 'startFlappy') {
+		game.flappyState = {
+			controllingLibUser: '',
+			controllingFascistUser: '',
+			liberalScore: 0,
+			fascistScore: 0,
+			pylonDensity: 1.3,
+			flapDistance: 1,
+			pylonOffset: 1.3,
+			passedPylonCount: 0
+		};
+
+		game.general.status = 'FLAPPY HITLER: 0 - 0';
+		io.sockets.in(game.general.uid).emit('gameUpdate', game);
+
+		game.flappyState.pylonGenerator = setInterval(() => {
+			const offset = Math.floor(Math.random() * 50 * game.flappyState.pylonOffset);
+			const newData = {
+				type: 'newPylon',
+				pylonType: 'normal',
+				offset
+			};
+
+			updateFlappyRoom(newData);
+		}, 1500 * game.flappyState.pylonDensity);
+	}
+
+	if (data.type === 'collision') {
+		game.flappyState[`${data.team}Score`]++;
+		clearInterval(game.flappyState.pylonGenerator);
+		// game.general.status = 'FLAPPY HITLER: x - x';
+		// io.sockets.in(game.general.uid).emit('gameUpdate', game);
+	}
+
+	if (data.type === 'passedPylon') {
+		game.flappyState.passedPylonCount++;
+		game.general.status = `FLAPPY HITLER: ${game.flappyState.liberalScore} - ${game.flappyState.fascistScore} (${game.flappyState.passedPylonCount})`;
+
+		io.sockets.in(game.general.uid).emit('gameUpdate', game);
+	}
+};
